add new components for playlist creation
