+++ src/components/UserProfile.js
@@ -1,21 +1,14 @@
 import React, { useState, useEffect } from "react";
+import axios from 'axios';
+
 
 function UserProfile(props){
-  const { accessToken } = props
-
-  useEffect(() => {
-    axios.get('https://api.spotify.com/v1/me', {
-      headers: {
-        'Authorization': {accessToken}
-      }
-    })
-    .then(response => {
-      console.log(response)
-    })
-  })
+  const [userData, setUserData] = useState({})
 
   return (
-
+    <div>
+      <h2>Signed In As {userData.display_name}</h2>
+    </div>
   )
 }
